+++ src/fkooman/Rest/Service.php
@@ -26,7 +26,10 @@ use fkooman\Http\Exception\InternalServerErrorException;
 use fkooman\Http\Exception\HttpException;
 use fkooman\Http\Exception\MethodNotAllowedException;
 use fkooman\Http\Exception\NotFoundException;
-use fkooman\Rest\Exception\ServiceException;
+use InvalidArgumentException;
+use RuntimeException;
+use LogicException;
+use BadFunctionCallException;
 
 class Service
 {
@@ -58,7 +61,7 @@ class Service
     public function setDefaultRoute($defaultRoute)
     {
         if (0 !== strpos($defaultRoute, '/')) {
-            throw new ServiceException('default route needs to start with a /');
+            throw new InvalidArgumentException('default route needs to start with a /');
         }
         $this->defaultRoute = $defaultRoute;
     }
@@ -196,7 +199,7 @@ class Service
                         return $response;
                     }
                     if (!is_string($response)) {
-                        throw new ServiceException("unsupported callback return value");
+                        throw new RuntimeException("unsupported callback return value");
                     }
                     $responseObj = new Response();
                     $responseObj->setContent($response);
@@ -250,7 +253,7 @@ class Service
         // check for variables in the requestPattern
         $pma = preg_match_all('#:([\w]+)\+?#', $requestPattern, $matches);
         if (false === $pma) {
-            throw new ServiceException("regex for variable search failed");
+            throw new LogicException("regex for variable search failed");
         }
         if (0 === $pma) {
             // no variables in the pattern, pattern and request must be identical
@@ -272,7 +275,7 @@ class Service
         $parameters = array();
         $pm = preg_match("#^".$requestPattern."$#", $request->getPathInfo(), $parameters);
         if (false === $pm) {
-            throw new ServiceException("regex for path matching failed");
+            throw new LogicException("regex for path matching failed");
         }
         if (0 === $pm) {
             // request path does not match pattern
@@ -316,7 +319,7 @@ class Service
                     // object
                     if (!array_key_exists($p->getClass()->getName(), $paramsAvailableForCallback)) {
                         if (!$p->isDefaultValueAvailable()) {
-                            throw new ServiceException("parameter expected by callback not available");
+                            throw new BadFunctionCallException("parameter expected by callback not available");
                         }
                     } else {
                         $cbParams[] = $paramsAvailableForCallback[$p->getClass()->getName()];
@@ -325,7 +328,7 @@ class Service
                     // internal type
                     if (!array_key_exists($p->getName(), $paramsAvailableForCallback)) {
                         if (!$p->isDefaultValueAvailable()) {
-                            throw new ServiceException("parameter expected by callback not available");
+                            throw new BadFunctionCallException("parameter expected by callback not available");
                         }
                     } else {
                         $cbParams[] = $paramsAvailableForCallback[$p->getName()];